Fixed a problem updating a translation request
updateTranslationRequest always tried to set a status - either DRAFT or
SUBMITTED. Because a user is allowed to update metadata field even a TR
is not in DRAFT status, new status in the change set should be null
unless isSubmit() is called for the change set.

Provided a test case checking if metadata field can be modified even a
TR is not in DRAFT status.

Fixes #30.